requirements.txt
streamlit>=1.35.0
yfinance>=0.2.40
pandas>=2.2.0
numpy>=1.26.0
plotly>=5.22.0
matplotlib>=3.8.0
fpdf2>=2.7.0
kaleido>=0.2.1
